[refactor] Swagger setup
